+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@ General:
    - Unified naming of commands and parameters (`layer` is `model` now, ...)
    - Added VoxelBuilder (`vbx`) support
    - Fixed missing group transform handling for MagicaVoxel files
+   - Added support for MagicaVoxel `xraw` format
 
 VoxConvert: